requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.3.1
async-timeout==3.0.1
attrs==23.2.0
chardet==4.0.0
discord.py==2.3.2
frozenlist==1.4.1
idna==3.6
iniconfig==2.0.0
multidict==6.0.5
mypy==1.4.0
mypy-extensions==1.0.0
packaging==23.2
pluggy==1.4.0
pytest==8.0.0
python-dotenv==1.0.1
typing-extensions==4.8.0
yarl==1.9.4
pytest-subtests==0.11.0
pytest-asyncio==0.22.0
flake8==6.1.0
pycodestyle==2.11.0
black==23.12.0
